% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aphia_latin}
\alias{aphia_latin}
\title{Latin names for aphia code}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2011 rows and 2 columns.
}
\usage{
aphia_latin
}
\description{
A data frame with 2011 rows and 2 columns:
\describe{
  \item{aphia}{aphia code}
  \item{latin}{Latin name}
}
}
\keyword{datasets}
