% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{asfis}
\alias{asfis}
\title{FAO species codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13417 rows and 5 columns.
}
\source{
<https://www.fao.org/fishery/static/ASFIS/ASFIS_sp.zip>
}
\usage{
asfis
}
\description{
A data frame with 13417 rows and 5 columns:
\describe{
  \item{species}{FAO 3-letter code}
  \item{latin}{Latin name}
  \item{english_name}{English name}
  \item{family}{Family}
  \item{order}{Order}
}
}
\keyword{datasets}
