% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpue_per_length_per_haul.R
\name{cpue_per_length_per_haul}
\alias{cpue_per_length_per_haul}
\title{cpue_per_length_per_haul}
\usage{
cpue_per_length_per_haul(hh, hl, tidy = TRUE)
}
\arguments{
\item{hh}{datras haul dataframe}

\item{hl}{datras length frequency dataframe}

\item{tidy}{If FALSE, first tidy the dataframes using functions tidy_hh and
tidy_hl (not implemented)}
}
\description{
Here put text on what is done
}
