% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_datras.R
\name{dr_calccpue}
\alias{dr_calccpue}
\title{Calculate with DATRAS data}
\usage{
dr_calccpue(d, hh)
}
\arguments{
\item{hh}{A tidy haul object obtained from tidyices::dr_getdata}
}
\value{
A tibble of the same dimention as the input
}
\description{
Calculate with DATRAS data
}
