% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_datras.R
\name{dr_calccpue_ca}
\alias{dr_calccpue_ca}
\title{dr_calccpue_ca}
\usage{
dr_calccpue_ca(d, hh)
}
\arguments{
\item{d}{A tidy ca object obtained from tidydatras::dr_getdata}

\item{hh}{A tidy haul object obtained from tidydatras::dr_getdata}
}
\value{
A tibble
}
\description{
dr_calccpue_ca
}
