% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_datras.R
\name{dr_calccpue_hl}
\alias{dr_calccpue_hl}
\title{dr_calccpue_hl}
\usage{
dr_calccpue_hl(d, hh)
}
\arguments{
\item{d}{A tidy hl object obtained from tidyices::dr_getdata}

\item{hh}{A tidy haul object obtained from tidyices::dr_getdata}
}
\value{
A tibble
}
\description{
dr_calccpue_hl
}
