% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_calclengthindex.R
\name{dr_calclengthindex}
\alias{dr_calclengthindex}
\title{Download, tidy and process length data into index (including zero hauls)}
\usage{
dr_calclengthindex(
  surveys = "FR-CGFS",
  years = 2005:2022,
  quarters = 4,
  fromweb = FALSE,
  folder = "C:/DATA/DATRAS/raw",
  quiet = TRUE
)
}
\arguments{
\item{surveys}{like "NS-IBTS"}

\item{years}{like "2000:2022"}

\item{quarters}{like c(1,3)}

\item{fromweb}{whether to get the data from Datras website (default is TRUE).}

\item{folder}{folder name if data is to be retrieved from existing file.}

\item{quiet}{whether to prevent messages from being shown on screen (default is TRUE).}
}
\value{
a tibble
}
\description{
Download, tidy and process length data into index (including zero hauls)
}
