% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doodle.R
\name{dr_doodle}
\alias{dr_doodle}
\title{Download and tidy length data}
\usage{
dr_doodle(survey_quarter = "FR-CGFS_4", years = 1989:2022)
}
\arguments{
\item{survey_quarter}{like "NS-IBTS_3"}

\item{years}{like "2000:2022"}
}
\value{
a tibble
}
\description{
Download and tidy length data
}
