% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_download_all.R
\name{dr_download_all}
\alias{dr_download_all}
\title{Download all DATRAS data}
\usage{
dr_download_all(
  save_dir,
  surveys = "all",
  years = "all",
  quarters = "all",
  quiet = FALSE
)
}
\arguments{
\item{save_dir}{The path to the directory where data are to be stored}

\item{surveys}{survey name or names (optional)}

\item{years}{year selection (optional)}

\item{quarters}{quarter selection (optional)}

\item{quiet}{boolean (default FALSE) that regulates whether info is being shown}
}
\description{
Saves rds files of all surveys HH, HL and CA data into a specified directory.
This takes a while so suggest to run this function in an R-script that is run via
nohub or "RStudio Background Jobs".
}
