% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoinside.R
\name{dr_geoinside}
\alias{dr_geoinside}
\title{Title}
\usage{
dr_geoinside(lon, lat, s, variable)
}
\arguments{
\item{lon}{xxx}

\item{lat}{xxx}

\item{s}{xxx}

\item{variable}{xxx}
}
\value{
A vector
}
\description{
Title
}
