% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_getflexoverview.R
\name{dr_getflexoverview}
\alias{dr_getflexoverview}
\title{Get an overview of available years and quarter for flexfiles}
\usage{
dr_getflexoverview(surveys, years, quarters)
}
\arguments{
\item{surveys}{A character vector specifying the surveys of interest. E.g.
c("ROCKALL", "NS-IBTS").}
}
\value{
A tibble containing variable survey, year and quarter and flexfile (0 or 1)
}
\description{
Get an overview of available years and quarter for flexfiles
}
