% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datras.R
\name{dr_getoverview}
\alias{dr_getoverview}
\title{Get an overview of avialable years and quarter}
\usage{
dr_getoverview(surveys)
}
\arguments{
\item{surveys}{A character vector specifying the surveys of interest. E.g.
c("ROCKALL", "NS-IBTS").}
}
\value{
A tibble containing variable survey, year and quarter
}
\description{
Get an overview of avialable years and quarter
}
