% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_generator.R
\name{dr_idseparate}
\alias{dr_idseparate}
\title{dr_idseparate}
\usage{
dr_idseparate(d, remove = TRUE)
}
\arguments{
\item{d}{A dataframe containing variable id}

\item{remove}{If TRUE, remove input column from output data frame.}
}
\description{
Separate the haul id code into its components
}
