% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_generator.R
\name{dr_idunite}
\alias{dr_idunite}
\title{dr_idunite}
\usage{
dr_idunite(d, remove = TRUE)
}
\arguments{
\item{d}{A dataframe containing variable id}

\item{remove}{If TRUE, remove input columns from output data frame.}
}
\description{
Unite varible that constitute a haul identification (year, quarter, ship, gear,
haulno) into variable id
}
