% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_settypes.R
\name{dr_settypes}
\alias{dr_settypes}
\title{Set the datras data types}
\usage{
dr_settypes(d)
}
\arguments{
\item{d}{a datras tibble}
}
\value{
A tibble containing tibble with appropriate data types
}
\description{
Set the datras data types
}
