% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_datras.R
\name{dr_tidy}
\alias{dr_tidy}
\title{Tidy DATRAS data}
\usage{
dr_tidy(d)
}
\arguments{
\item{d}{An object obtained from tidydatras::dr_getdata or icesDatras::getDATRAS}
}
\value{
A tibble of the same dimention as the input
}
\description{
Tidy DATRAS data
}
