% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_datras.R
\name{dr_tidyhh}
\alias{dr_tidyhh}
\title{dr_tidyhh}
\usage{
dr_tidyhh(d)
}
\arguments{
\item{d}{DATRAS raw haul dataframe, as obtained via the function tidyices::}
}
\value{
a tibble
}
\description{
dr_tidyhh
}
