% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_datras.R
\name{dr_tidyhl}
\alias{dr_tidyhl}
\title{dr_tidyhl}
\usage{
dr_tidyhl(d)
}
\arguments{
\item{d}{DATRAS raw length dataframe, as obtained via the function
icesDatras::getDATRAS}
}
\value{
A tibble
}
\description{
dr_tidyhl
}
