% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrs_overview.R
\name{dtrs_overview}
\alias{dtrs_overview}
\title{Get an overview of avialable years and quarter}
\usage{
dtrs_overview(survey)
}
\arguments{
\item{survey}{A character vector specifying the surveys of interest. E.g.
c("ROCKALL", "NS-IBTS").}
}
\value{
A tibble containing variable survey, year and quarter
}
\description{
Get an overview of avialable years and quarter
}
