% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrs_tidy.R
\name{dtrs_tidy}
\alias{dtrs_tidy}
\title{Tidy DATRAS data}
\usage{
dtrs_tidy(d, all_variables = FALSE, ...)
}
\arguments{
\item{d}{Any nontidied table of the form HH, HL or CA}

\item{all_variables}{A boolean (default FALSE) indicating
if all or only "necessary" variables are return.}

\item{...}{}
}
\description{
NEED GOOD EXPLANATION HERE
}
