% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrs_tidy_hh.R
\name{dtrs_tidy_hh}
\alias{dtrs_tidy_hh}
\title{dtrs_tidy_hh}
\usage{
dtrs_tidy_hh(df, all_variables = FALSE)
}
\arguments{
\item{df}{A datras haul dataframe}

\item{all_variables}{A TRUE/FALSE flag. Should all variables
(default FALSE) or all variable (TRUE) be returned}
}
\value{
data.frame
}
\description{
dtrs_tidy_hh
}
