% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo_inside.R
\name{geo_inside}
\alias{geo_inside}
\title{Spatial location of datapoints}
\usage{
geo_inside(lon, lat, map, variable)
}
\arguments{
\item{lon}{A numerical vector}

\item{lat}{A numerical vector}

\item{map}{Normally a spatialPolygonDataFrame (sf or sp)}

\item{variable}{The variable name, stored in the map attribute table to be
returned. If missing (default) only boolean vector is returned indicating
if coordinates are inside or outside any region.}
}
\description{
Returns spatial attributes of coordinates (longitude and latitude)
given spatial polygons.
}
