% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latin.R
\name{get_latin}
\alias{get_latin}
\title{get_latin}
\usage{
get_latin(df)
}
\arguments{
\item{df}{A dataframe that contains a columname "validaphia"}
}
\value{
A datafram with columns validaphia, latin and species name
}
\description{
get_latin
}
