% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ibts_rect}
\alias{ibts_rect}
\title{IBTS rect data}
\format{
An object of class \code{NULL} of length 0.
}
\source{
<https://www.ices.dk/data/Documents/DATRAS/Indices_Calculation_Steps_IBTS.pdf>
}
\usage{
ibts_rect
}
\description{
A data frame with 211 rows and 2 columns:
\describe{
  \item{statrec}{ICES Statistical Rectangle}
  \item{weight}{Weight of rectangle in index calculation}
}
}
\keyword{datasets}
