% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_separate.R
\name{id_separate}
\alias{id_separate}
\title{id_separatate}
\usage{
id_separate(d, remove = TRUE)
}
\arguments{
\item{d}{A dataframe containing variable id}

\item{remove}{If TRUE, remove input column from output data frame.}
}
\description{
Separate the haul id code into its components
}
