% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_unite.R
\name{id_unite}
\alias{id_unite}
\title{id_unite}
\usage{
id_unite(d, remove = TRUE)
}
\arguments{
\item{d}{A dataframe containing variable id}

\item{remove}{If TRUE, remove input columns from output data frame.}
}
\description{
Unite varible that constitute a haul identification (year, quarter, ship, gear,
haulno) into variable id
}
