% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ns_ibts_rf}
\alias{ns_ibts_rf}
\title{North Sea IBTS FS area}
\format{
An object of class \code{sf} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10 rows and 9 columns.
}
\source{
<https://www.ices.dk/data/Documents/DATRAS/DATRAS_Field_descriptions_and_example_file_May2022.xlsx>
}
\usage{
ns_ibts_rf
}
\description{
A shapefile, polygons often used in ALK compilations
}
\keyword{datasets}
