% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_exchange.R
\name{read_exchange}
\alias{read_exchange}
\title{read_exchange}
\usage{
read_exchange(fname)
}
\arguments{
\item{fname}{The name of the exchange file (usually a csv-file)}
}
\value{
a list
}
\description{
read_exchange
}
