% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reco}
\alias{reco}
\title{RECO vessel codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12962 rows and 2 columns.
}
\source{
<https://vocab.ices.dk/?ref=315>
}
\usage{
reco
}
\description{
A data frame with 12962 rows and 2 columns:
\describe{
  \item{code}{Vessel code}
  \item{vessel}{Vessel name}
}
}
\keyword{datasets}
