% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_raw.R
\name{tidy_hh}
\alias{tidy_hh}
\title{tidy_hh}
\usage{
tidy_hh(d, remove = TRUE)
}
\arguments{
\item{d}{DATRAS raw haul dataframe, as obtained via the function
icesDatras::getDATRAS}

\item{remove}{logical; remove all but "essential" columns. Currently not active.}
}
\value{
a tibble
}
\description{
tidy_hh
}
