% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_raw.R
\name{tidy_hl}
\alias{tidy_hl}
\title{tidy_hl}
\usage{
tidy_hl(d, remove = TRUE)
}
\arguments{
\item{d}{DATRAS raw length dataframe, as obtained via the function
icesDatras::getDATRAS}

\item{remove}{logical; remove all but "essential" columns.  Currently not active.}
}
\value{
A tibble
}
\description{
tidy_hl
}
