% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lengths.R
\name{tidy_lengths}
\alias{tidy_lengths}
\title{tidy_lengths}
\usage{
tidy_lengths(hl, hh, species)
}
\arguments{
\item{hl}{datras untidy length dataframe, as obtained via the function
icesDatras::getDATRAS}

\item{hh}{datras tidy haul data}

\item{species}{dataframe with species code, of the form obtained via function
get_species. Required column names are valid_aphia and latin. If dataframe
not supplied in the function argument, it will be automatically obtained.}
}
\value{
A tibble with 6 variables:

\describe{
  \item{id}{haul id}
  \item{latin}{Latin name of species}
  \item{species}{English name of species}
  \item{sex}{Sex of fish}
  \item{length}{Length in centimeters}
  \item{n}{Number of fish per standardized to 60 minute haul}
}
}
\description{
tidy_lengths
}
