% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datras_field_types}
\alias{datras_field_types}
\title{Datras field types}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 125 rows and 3 columns.
}
\source{
<https://www.ices.dk/data/Documents/DATRAS/DATRAS_Field_descriptions_and_example_file_May2022.xlsx>
}
\usage{
datras_field_types
}
\description{
A table containing the field types of the datras variables
}
\details{
A data frame with 125 rows and 3 columns:
\describe{
  \item{field}{DATRAS variable name as returned by icesDatras::getDATRAS}
  \item{type}{The value type}
  \item{record}{The DATRAS data type - "HH": haul data, "HL": length-based data, "CA": age-based data}
}
}
\keyword{datasets}
