% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_calccpue.R
\name{dr_calccpue}
\alias{dr_calccpue}
\title{Calculate with DATRAS data}
\usage{
dr_calccpue(d, hh, fl)
}
\arguments{
\item{d}{A tidy hl or ca object obtained from tidydatras::dr_getdata}

\item{hh}{A tidy haul object obtained from tidydatras::dr_getdata}
}
\value{
A tibble of the same dimension as the input
}
\description{
Calculate with DATRAS data
}
