% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_datras.R
\name{dr_calccpue_ca}
\alias{dr_calccpue_ca}
\title{dr_calccpue_ca}
\usage{
dr_calccpue_ca(d, hh)
}
\arguments{
\item{d}{DATRAS raw length dataframe, as obtained via the function
icesDatras::getDATRAS}
}
\value{
A tibble
}
\description{
dr_calccpue_ca
}
