% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_calccpue.R
\name{dr_calccpue_hl}
\alias{dr_calccpue_hl}
\title{dr_calccpue_hl}
\usage{
dr_calccpue_hl(d, hh, fl)
}
\arguments{
\item{d}{A tidy hl object obtained from tidydatras::dr_getdata}

\item{hh}{A tidy haul object obtained from tidydatras::dr_getdata}
}
\value{
A tibble
}
\description{
dr_calccpue_hl
}
