% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_all.R
\name{dr_download_all}
\alias{dr_download_all}
\title{Download all DATRAS data}
\usage{
dr_download_all(save_dir, verbose = TRUE)
}
\arguments{
\item{save_dir}{The path to the directory where data are to be stored}

\item{verbose}{boolean (default TRUE) that prints out survey being worked on}
}
\description{
Saves rds files of all surveys HH, HL and CA data into a specified directory.
This takes a while so suggest to run this function in an R-script that is run via
nohub or "RStudio Background Jobs".
}
