% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_datras.R
\name{dr_getdata}
\alias{dr_getdata}
\title{Get any DATRAS data}
\usage{
dr_getdata(record = "HH", surveys, years, quarters, quiet = TRUE)
}
\arguments{
\item{record}{the data type required: "HH" haul data, "HL" length-based data, "CA" age-based data.}

\item{surveys}{the survey acronym e.g. NS-IBTS}

\item{years}{a vector of years of the survey, e.g. c(2010, 2012) or 2005:2010.}

\item{quarters}{a vector of quarters of the year the survey took place, i.e. c(1, 4) or 1:4.}

\item{quiet}{whether to prevent messages from being shown on screen (default is TRUE).}
}
\value{
A tibble
}
\description{
This a wrapper around icesDatras::getDATRAS that ensure that proper variable
type is returned.
}
