% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latin.R
\name{get_latin}
\alias{get_latin}
\title{get_latin}
\usage{
get_latin(df)
}
\arguments{
\item{df}{A dataframe that contains a columname "validaphia"}
}
\value{
A dataframe with columns validaphia, latin and english species name.
If no object is passed, then read species code from icesVocab
}
\description{
Accesss ICES getSpecies webserver. Equivalent to what one would obtain using
https://datras.ices.dk/Data_products/qryspec.aspx
}
