% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_raw.R
\name{tidy_ca}
\alias{tidy_ca}
\title{tidy_ca}
\usage{
tidy_ca(d, remove = TRUE)
}
\arguments{
\item{d}{DATRAS raw age dataframe, as obtained via the function
icesDatras::getDATRAS}

\item{remove}{logical; remove all but "essential" columns. Currently not active.}
}
\value{
A tibble with

\describe{
  \item{id}{haul id}
  \item{latin}{Latin name of species}
  \item{species}{English name of species}
  \item{length}{Length in centimeters}
  \item{sex}{Sex of fish}
  \item{maturity}{The maturity scale}
  \item{age}{Age in years}
  \item{wgt}{Weight in grammes}
  \item{n}{Number of fish}
}
}
\description{
tidy_ca
}
