% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_hh.R
\name{tidy_hh}
\alias{tidy_hh}
\title{tidy_hh}
\usage{
tidy_hh(df, all_variables = FALSE)
}
\arguments{
\item{df}{A datras haul dataframe}

\item{all_variables}{A TRUE/FALSE flag. Should all variables
(default FALSE) or all variable (TRUE) be returned}
}
\value{
data.frame
}
\description{
tidy_hh
}
