% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lengts.R
\name{tidy_lengths}
\alias{tidy_lengths}
\title{tidy_lengths}
\usage{
tidy_lengths(hl, hh, species)
}
\arguments{
\item{hl}{datras length untidy length data}

\item{hh}{tidy datras haul data}

\item{species}{species code}
}
\description{
tidy_lengths
}
